Refactor: storybook specification에서 외부 테스트 파일 로드하도록 변경
테스트 파일 위치 이동
